# Conflicts:
#	.idea/workspace.xml
#	composer.json
#	composer.lock
#	coverage/BankModel.php.html
#	coverage/BaseModel.php.html
#	coverage/FactoryPattern.php.html
#	coverage/UserModel.php.html
#	coverage/dashboard.html
#	coverage/index.html
#	form_user.php
#	list_banks.php
#	models/BankModel.php
#	models/BaseModel.php
#	models/UserModel.php
#	phpunit.xml
#	public/css/styles.css
#	test.php
#	tests/UserModelTest.php
#	views/header.php

+++ models/UserModel.php
@@ -1,20 +1,30 @@
 <?php
 
-use function PHPSTORM_META\type;
-
 require_once 'BaseModel.php';
 
-class UserModel extends BaseModel {
+class UserModel extends BaseModel
+{
+    public static function getInstance()
+    {
+        if(self::$_instance !== null){
+            return self::$_instance;
+        }
+        self::$_instance = new self();
+        return self::$_instance;
+    }
 
-    public function findUserById($id) {
-        $sql = 'SELECT * FROM users WHERE id = '.$id;
+    public function findUserById($id)
+    {
+        
+        $sql = 'SELECT * FROM users WHERE id = ' . $id;
         $user = $this->select($sql);
 
         return $user;
     }
 
-    public function findUser($keyword) {
-        $sql = 'SELECT * FROM users WHERE user_name LIKE %'.$keyword.'%'. ' OR user_email LIKE %'.$keyword.'%';
+    public function findUser($keyword)
+    {
+        $sql = 'SELECT * FROM users WHERE user_name LIKE %' . $keyword . '%' . ' OR user_email LIKE %' . $keyword . '%';
         $user = $this->select($sql);
 
         return $user;
@@ -22,13 +32,14 @@ class UserModel extends BaseModel {
 
     /**
      * Authentication user
-     * @param $userName
+     * @param $
      * @param $password
      * @return array
      */
-    public function auth($userName, $password) {
+    public function auth($userName, $password)
+    {
         $md5Password = md5($password);
-        $sql = 'SELECT * FROM users WHERE name = "' . $userName . '" AND password = "'.$md5Password.'"';
+        $sql = 'SELECT * FROM users WHERE name = "' . $userName . '" AND password = "' . $md5Password . '"';
 
         $user = $this->select($sql);
         return $user;
@@ -39,35 +50,61 @@ class UserModel extends BaseModel {
      * @param $id
      * @return mixed
      */
-    public function deleteUserById($id) {
-        $sql = 'DELETE FROM users WHERE id = '.$id;
-        return $this->delete($sql);
-
+    public function deleteUserById($id)
+    {   
+        $isAuth = $this->getUsers();
+        foreach ($isAuth as $item) {
+            if (md5($item['id']) == $id) {
+                $sql = 'DELETE FROM users WHERE id = ' . $item['id'];
+                return $this->delete($sql);
+            }
+        }
     }
+    // Delete user by id : Step 2
+    public function dropUserById($id)
+    {   
+        $sql = 'DELETE FROM users WHERE id = ' . $id;
+        return $this->delete($sql);
+    }
+    /**
+     * Delete user by id
+     * @param $id
+     * @return mixed
+     */
+    // public function deleteUserById($id)
+    // {
+    //     $sql = 'DELETE FROM users WHERE id = ' . $id;
+    //     return $this->delete($sql);
+    // }
 
     /**
      * Update user
      * @param $input
      * @return mixed
      */
-    public function updateUser($input) {
+    public function updateUser($input)
+    {
         $sql = 'UPDATE users SET 
-                 name = "' . $input['name'] .'", 
-                 password="'. md5($input['password']) .'"
+                 name = "' . mysqli_real_escape_string(self::$_connection, $input['name']) . '"
+                ,`fullname`="' . $input['full-name'] . '"
+                ,email="' . $input['email'] . '"
+                ,type="' . $input['type'] . '"
+                ,password="' . md5($input['password']) . '"
                 WHERE id = ' . $input['id'];
         $user = $this->update($sql);
-
         return $user;
     }
+
 
     /**
      * Insert user
      * @param $input
      * @return mixed
      */
-    public function insertUser($input) {
-        $sql = "INSERT INTO `app_web1`.`users` (`name`, `password`) VALUES (" .
-                "'" . $input['name'] . "', '".md5($input['password'])."')";
+    public function insertUser($input)
+    {
+        $sql = "INSERT INTO `app_web1`.`users` (`name`,`fullname`, `email`, `type`, `password`) VALUES (" .
+            "'" . $input['name'] . "', '" . $input['full-name'] . "' , '" . $input['email'] . "', '" . $input['type'] . "', '" . md5($input['password']) . "')";
 
         $user = $this->insert($sql);
 
@@ -79,10 +116,11 @@ class UserModel extends BaseModel {
      * @param array $params
      * @return array
      */
-    public function getUsers($params = []) {
+    public function getUsers($params = [])
+    {
         //Keyword
         if (!empty($params['keyword'])) {
-            $sql = 'SELECT * FROM users WHERE name LIKE "%' . $params['keyword'] .'%"';
+            $sql = 'SELECT * FROM users WHERE name LIKE "%' . $params['keyword'] . '%"';
 
             //Keep this line to use Sql Injection
             //Don't change
@@ -95,32 +133,4 @@ class UserModel extends BaseModel {
 
         return $users;
     }
-    public static function getInstance(){
-        if (self::$_instance !== null) {
-            return self::$_instance;
-        }
-        self::$_instance = new self();
-        return self::$_instance;
-    }
-    /**
-     * Convert to number
-     * @param $input
-     */
-    public function convertNumber($input):int {
-        if (is_string($input)) {
-            $input = (int)$input;
-        }
-        return $input;
-    }
-    /**
-     * For testing
-     * @param $a
-     * @param $b
-     */
-    public function sumb(int $a, int $b) {
-        $this->convertNumber($a);
-        $this->convertNumber($b);
-        return $a + $b;
-    }
-   
 }
